Separate downloading stuff from remote
