+++ pyproject.toml
@@ -61,6 +61,7 @@ pdbpp = "^0.10.3"
 
 [tool.poetry.scripts]
 nnu = "nnusf.cli:command"
+nns = "nnusf.get:command"
 
 [tool.poetry-dynamic-versioning]
 enable = true
