+++ src/nnusf/get/base.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+import click
+
+CONTEXT_SETTINGS = dict(help_option_names=["-h", "--help"])
+
+
+@click.group(context_settings=CONTEXT_SETTINGS)
+def command():
+    pass
