+++ src/nnusf/get/log.py
@@ -1,0 +1,11 @@
+# -*- coding: utf-8 -*-
+import logging
+
+from rich.logging import RichHandler
+
+logging.basicConfig(
+    level=logging.INFO,
+    format="%(message)s",
+    datefmt="[%X]",
+    handlers=[RichHandler()],
+)
